requirements (1).txt
streamlit_option_menu == 0.3.2
requests==2.28.1
streamlit-lottie==0.0.3
times==0.7
htbuilder==0.6.1
transformers==4.29.2
torch==2.0.1
